fix(lift): widen vector lane offset arithmetic

+++ lib/lift/AArch64/AArch64LiftArith.cpp
@@ -81,8 +81,10 @@ bool liftArith(AArch64Lifter &L, AArch64Lifter::LiftState &S,
         for (unsigned I = 0; I < NLanes; ++I) {
           NdVar La = S.makeTemp(LaneSz);
           NdVar Lb = S.makeTemp(LaneSz);
-          S.emit(NdOp::SUBBYTES, La, {A, NdVar::cst(I * LaneSz, 4)});
-          S.emit(NdOp::SUBBYTES, Lb, {B, NdVar::cst(I * LaneSz, 4)});
+          S.emit(NdOp::SUBBYTES, La,
+                 {A, NdVar::cst(static_cast<uint64_t>(I) * LaneSz, 4)});
+          S.emit(NdOp::SUBBYTES, Lb,
+                 {B, NdVar::cst(static_cast<uint64_t>(I) * LaneSz, 4)});
           NdVar Lr = S.makeTemp(LaneSz);
           S.emit(NdOp::INT_ADD, Lr, {La, Lb});
           if (I == 0) {
